Fix integer-overflow and buffer over-read bugs across format parsers
- lib-gfa.c: widen ScanGFA's header offset check to 64-bit to avoid
  wraparound with attacker-controlled offsets near UINT32_MAX
- lib-bea.c: widen dict node-count arithmetic to 64-bit to avoid the
  same wraparound zeroing the allocated table size
- lib-rnc.c: clamp negative remaining-length values before memcpy in
  the block refill, and reject match offsets that would read before
  the decode buffer
- lib-ash0.c: check CALLOC result for NULL before use
- lib-bars.c, lib-gar.c, lib-mkgpdxpac.c, lib-ndcubepac.c, lib-nud.c,
  lib-nxarc.c, lib-pac.c, lib-pgdat.c, lib-pkz.c, lib-tmpk.c: clamp
  strnlen()'s length cap to the bytes actually remaining in the
  loaded buffer instead of a fixed field size, fixing an out-of-bounds
  heap read for non-terminated embedded names near the end of a file

+++ project/src/lib-gfa.c
@@ -331,7 +331,7 @@ enumError ScanGFA (gfa_t *gfa, const u8 *data, uint size)
 	const u32 data_off = rd_le32 (data + 0x14);
 	const u32 data_size = rd_le32 (data + 0x18);
 
-	if (info_off + 4 > size || data_off + 16 > size)
+	if ((u64)info_off + 4 > size || (u64)data_off + 16 > size)
 		return EINVAL;
 	if ((u64)data_off + data_size > size)
 		return EINVAL;
